requirements.txt
async-timeout==4.0.2
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
packaging==21.3
passlib==1.7.4
pymongo==4.2.0
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
